backend/requirements.txt
# Core Django（5.1 开始对新解释器更友好）
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4

# Database
psycopg2-binary>=2.9.9

# Cache
redis>=5.0,<5.1
django-redis>=5.4,<5.5

# Storage
django-storages>=1.14,<1.15
boto3>=1.34,<=1.36

# Email
django-anymail>=10.2,<10.3

# CORS
django-cors-headers>=4.3,<4.4

# Env
python-dotenv>=1.0,<1.1

# API Docs
drf-spectacular>=0.27,<0.28

# Utilities
django-filter>=23.5,<24.0
Pillow>=11.0.0  # 关键：换到支持 3.13 的版本

# Prod
gunicorn>=21.2,<21.3

# Dev
ipython>=8.18,<8.19
